+++ react/components/enhancers/index.js
@@ -1,0 +1,3 @@
+export { default as enhance } from './enhance'
+export { default as withShape } from './withShape'
+export { default as withOrientation } from './withOrientation'